# Conflicts:
#	Server/src/test/java/com/example/nyurates/ProfessorServiceTests.java

+++ Server/src/main/java/com/example/nyurates/controller/AdminController.java
@@ -3,8 +3,10 @@ package com.example.nyurates.controller;
 import com.example.nyurates.entity.Admin;
 import com.example.nyurates.entity.results.*;
 import com.example.nyurates.service.AdminService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.support.SessionStatus;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -40,8 +42,19 @@ public class AdminController {
      * @return Result
      */
     @GetMapping (value = "/stats")
-    public Result getStats(){
-        return adminService.getStatistics();    
+    public Result getStats(HttpServletRequest request){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.getStatistics();
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
     }
 
     /**
@@ -50,8 +63,20 @@ public class AdminController {
      * @return StudentListResult
      */
     @PostMapping(value = "/student_list")
-    public StudentListResult studentList(@RequestBody Map<String, String> params){
-        return adminService.studentList(params.get("name"), params.get("netid"), params.get("email"));
+    public Result studentList(HttpServletRequest request, @RequestBody Map<String, String> params){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.studentList(params.get("name"), params.get("netid"), params.get("email"));
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
+
     }
 
     /**
@@ -60,8 +85,19 @@ public class AdminController {
      * @return ProfListResult
      */
     @PostMapping(value = "/prof_list")
-    public ProfListResult profList(@RequestBody Map<String, String> params){
-        return adminService.profList(params.get("name"), params.get("netid"), params.get("email"), params.get("department"));
+    public Result profList(HttpServletRequest request, @RequestBody Map<String, String> params){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.profList(params.get("name"), params.get("netid"), params.get("email"), params.get("department"));
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
     }
 
     /**
@@ -71,9 +107,19 @@ public class AdminController {
      * @return ReportListResult
      */
     @RequestMapping(value = "/getreports")
-    public ReportListResult getReports(HttpSession session, @RequestBody Map<String, Object> params){
-        System.out.println(session.getAttribute("role"));
-        return adminService.getReports((Long) params.get("report_id"), (Long) params.get("comment_id"), (String) params.get("comment_user"), (String) params.get("course_code"));
+    public Result getReports(HttpServletRequest request, @RequestBody Map<String, Object> params){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.getReports((Long) params.get("report_id"), (Long) params.get("comment_id"), (String) params.get("comment_user"), (String) params.get("course_code"));
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
     }
 
     /**
@@ -82,8 +128,19 @@ public class AdminController {
      * @return Result
      */
     @PostMapping(value = "/reviewcomment")
-    public Result reviewComment(@RequestBody Map<String, Object> params) {
-        return adminService.reviewComment((Integer) params.get("comment_id"), (Integer) params.get("report_id"), (Boolean) params.get("validity"));
+    public Result reviewComment(HttpServletRequest request, @RequestBody Map<String, Object> params) {
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.reviewComment((Integer) params.get("comment_id"), (Integer) params.get("report_id"), (Boolean) params.get("validity"));
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
     }
 
     /**
@@ -92,9 +149,20 @@ public class AdminController {
      * @return ProfReqResult
      */
     @RequestMapping (value="/getprofrequests")
-    public ProfReqResult getProfRequests(@RequestBody Map<String, Object> params){
-        return adminService.getProfRequests();
-    }
+    public Result getProfRequests(HttpServletRequest request, @RequestBody Map<String, Object> params){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.getProfRequests();
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
+    } 
 
     /**
      * Handle Add Professor Request
@@ -102,8 +170,19 @@ public class AdminController {
      * @return Result
      */
     @RequestMapping (value="/handleprofrequests")
-    public Result handleProfRequests(@RequestBody Map<String, Object> params){
-        return adminService.handleProfReq((Integer) params.get("request_id"), (boolean) params.get("operation"));
+    public Result handleProfRequests(HttpServletRequest request, @RequestBody Map<String, Object> params){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.handleProfReq((Integer) params.get("request_id"), (boolean) params.get("operation"));
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
     }
 
     /**
@@ -112,8 +191,19 @@ public class AdminController {
      * @return Result
      */
     @RequestMapping (value="/deletestudent")
-    public Result deleteStudent(@RequestBody Map<String, String> params){
-        return adminService.deleteStudent(params.get("email"));
+    public Result deleteStudent(HttpServletRequest request, @RequestBody Map<String, String> params){
+        HttpSession session = request.getSession(false);
+        try{
+            if (((String) session.getAttribute("role")).equals("admin")){
+                return adminService.deleteStudent(params.get("email"));
+            }
+            else{
+                return new UnauthorizedResult();
+            }
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
     }
 
     /**
@@ -122,13 +212,41 @@ public class AdminController {
      * @return Result
      */
     @GetMapping (value="/validate")
-    public Result validateAdmin(HttpSession session){
+    public Result validateAdmin(HttpServletRequest request){
+        HttpSession session = request.getSession(false);
         Result result = new Result();
-        if(((String) session.getAttribute("role")).equals("admin")){
+        result.setCode(400);
+        try{
+            if(((String) session.getAttribute("role")).equals("admin")){
+                result.setCode(200);
+                result.setMsg("Validate Success");
+                return result;
+            }
+            return new UnauthorizedResult();
+        }
+        catch (Exception e){
+            return new UnauthorizedResult();
+        }
+    }
+
+    @GetMapping (value="/logout")
+    public Result adminLogout(HttpServletRequest request, SessionStatus sessionStatus){
+        HttpSession session = request.getSession(false);
+        try{
+            session.removeAttribute("role");
+            session.removeAttribute("email");
+            session.invalidate();
+            sessionStatus.setComplete();
+            Result result = new Result();
             result.setCode(200);
-            result.setMsg("Validate Success");
+            result.setMsg("Successfully logged out.");
             return result;
         }
-        return result;
+        catch (Exception e){
+            e.printStackTrace();
+            Result result = new Result();
+            result.setCode(400);
+            return result;
+        }
     }
 }
